pyproject.toml
[tool.poetry]
name = "neural-bet"
version = "0.1.0"
description = "Double-Blind AI Oracle for Football Analysis"
authors = ["Antigravity <architect@antigravity.ai>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.scripts]
# CLI entry points - run from anywhere after `poetry install`
neuralbet = "src.cli:main"
neuralbet-pipeline = "src.cli:run_pipeline"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.6.0"
pydantic-settings = "^2.2.0"
langchain = "^0.3.0"
langgraph = "^0.2.0"
langchain-google-genai = "^1.0.0"
langchain-groq = "^0.1.0"
pandas = "^2.2.0"
httpx = "^0.27.0"
beautifulsoup4 = "^4.12.0"
python-dotenv = "^1.0.0"
colorama = "^0.4.6"
tenacity = "^8.2.0"
soccerdata = "^1.8.0"  # FBRef scraping with anti-bot handling
textual = "^0.47.0"  # TUI framework

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.0.0"
isort = "^5.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
